Unify assignment patterns and literals
The ES6 spec actually states that object and array assignment patterns
are parsed this way: initially as literals, and then reparsed to make
sure that they are valid assignment patterns.

Since it is ok for this grammar to accept code that is semantically
invalid, we simply allow arbitrary literals as assignment patterns.
This removes a lot of ambiguities between expressions and assignment
patterns. :racehorse: