Remove the check for curly brace validity in external scanner
This fixes a parse error that started occurring in minified files
after #45. I thought that it was important not to produce an
automatic semicolon token if '}' was valid in order to properly
distinguish between objects and labelled statements for code
like this:

{
  a: b
}

But now that we have the `property_name` rule, there is already
an ambiguity in the parse stack in this situation, and the
automatic semicolon will only arise in the labelled statement
interpretation.